+++ R/aorsf-package.R
@@ -6,6 +6,7 @@
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start
 #' @importFrom Rcpp sourceCpp
+#' @importFrom collapse %==%
 #' @useDynLib aorsf, .registration = TRUE
 ## usethis namespace: end
 NULL