+++ R/collapse_misc.R
@@ -47,3 +47,5 @@ data_impute_nocheck <- function(data, cols, values){
  data
 
 }
+
+