+++ R/orsf_predict.R
@@ -2,8 +2,7 @@
 
 #' Compute predictions using ORSF
 #'
-#' Predicted risk or survival (someday also hazard or mortality)
-#'   from an ORSF model.
+#' Predicted risk, survival, hazard, or mortality from an ORSF model.
 #'
 #' @srrstats {G1.4} *documented with Roxygen*
 #' @srrstats {ML1.1} *using the terms 'train' and 'test'.*
@@ -37,6 +36,9 @@
 #'   - `r roxy_na_action_fail("new_data")`
 #'   - `r roxy_na_action_pass("new_data")`
 #'   - `r roxy_na_action_omit("new_data")`
+#'   - `r roxy_na_action_impute_meanmode('new_data')`. To clarify,
+#'     the mean and mode used to impute missing values are from the
+#'     training data of `object`, not from `new_data`.
 #'
 #' @param boundary_checks (_logical_) if `TRUE`, `pred_horizon` will be
 #'  checked to make sure the requested values are less than the maximum
@@ -97,7 +99,9 @@ predict.orsf_fit <- function(object,
  pred_horizon_order <- order(pred_horizon)
  pred_horizon_ordered <- pred_horizon[pred_horizon_order]
 
- cc <- which(stats::complete.cases(select_cols(new_data, names_x_data)))
+ cc <- which(
+  stats::complete.cases(select_cols(new_data, names_x_data))
+ )
 
  check_complete_cases(cc, na_action, nrow(new_data))
 
@@ -107,6 +111,8 @@ predict.orsf_fit <- function(object,
                           cols = get_names_x(object),
                           values = c(as.list(get_means(object)),
                                      as.list(get_modes(object))))
+
+  cc <- collapse::seq_row(new_data)
 
  }