+++ src/orsf.cpp
@@ -1428,15 +1428,15 @@ double lrt_multi(){
  vec_temp.resize( jit_vals.size() );
 
  // protection from going out of bounds with jit_vals(k) below
- if(j == 0) jit_vals(jit_vals.size()-1)--;
+ if(j == 0) jit_vals.at(jit_vals.size()-1)--;
 
  // put the indices of potential cut-points into vec_temp
  for(k = 0; k < vec_temp.size(); k++){
-  vec_temp[k] = XB(*(iit_best - jit_vals[k]));
+  vec_temp[k] = XB.at(*(iit_best - jit_vals[k]));
  }
 
  // back to how it was!
- if(j == 0) jit_vals(jit_vals.size()-1)++;
+ if(j == 0) jit_vals.at(jit_vals.size()-1)++;
 
  // if(verbose > 1){
  //
@@ -2293,24 +2293,24 @@ void new_pred_surv_multi(char pred_type){
 
   for(j = 0; j < times_pred.size(); j++){
 
-   time_pred = times_pred(j);
+   time_pred = times_pred.at(j);
 
-   if(time_pred < leaf_node(leaf_node.n_rows - 1, 0)){
+   if(time_pred < leaf_node.at(leaf_node.n_rows - 1, 0)){
 
     for(; i < leaf_node.n_rows; i++){
 
-     if (leaf_node(i, 0) > time_pred){
+     if (leaf_node.at(i, 0) > time_pred){
 
       if(i == 0)
        temp1 = pred_t0;
       else
-       temp1 = leaf_node(i-1, leaf_node_col);
+       temp1 = leaf_node.at(i-1, leaf_node_col);
 
       break;
 
-     } else if (leaf_node(i, 0) == time_pred){
+     } else if (leaf_node.at(i, 0) == time_pred){
 
-      temp1 = leaf_node(i, leaf_node_col);
+      temp1 = leaf_node.at(i, leaf_node_col);
       break;
 
      }
@@ -2320,11 +2320,11 @@ void new_pred_surv_multi(char pred_type){
    } else {
 
     // go here if prediction horizon > max time in current leaf.
-    temp1 = leaf_node(leaf_node.n_rows - 1, leaf_node_col);
+    temp1 = leaf_node.at(leaf_node.n_rows - 1, leaf_node_col);
 
    }
 
-   surv_pvec(j) = temp1;
+   surv_pvec.at(j) = temp1;
 
   }
 
@@ -2333,7 +2333,7 @@ void new_pred_surv_multi(char pred_type){
 
   if(iit < iit_vals.end()){
 
-   while(person_leaf == leaf_pred(*iit)){
+   while(person_leaf == leaf_pred.at(*iit)){
 
     surv_pmat.row(*iit) += surv_pvec.t();
     ++iit;
@@ -2427,7 +2427,7 @@ void new_pred_surv_uni(char pred_type){
   } else {
 
    // go here if prediction horizon > max time in current leaf.
-   temp1 = leaf_node(leaf_node.n_rows - 1, leaf_node_col);
+   temp1 = leaf_node.at(leaf_node.n_rows - 1, leaf_node_col);
 
    // --- EXPERIMENTAL ADD-ON --- //
    // if you are predicting beyond the max time in a node,
@@ -2441,14 +2441,14 @@ void new_pred_surv_uni(char pred_type){
 
   }
 
-  surv_pvec(*iit) += temp1;
+  surv_pvec.at(*iit) += temp1;
   ++iit;
 
   if(iit < iit_vals.end()){
 
-   while(person_leaf == leaf_pred(*iit)){
+   while(person_leaf == leaf_pred.at(*iit)){
 
-    surv_pvec(*iit) += temp1;
+    surv_pvec.at(*iit) += temp1;
     ++iit;
 
     if (iit == iit_vals.end()) break;
@@ -3160,7 +3160,8 @@ List orsf_fit(NumericMatrix& x,
               Function       f_beta,
               const char&    type_beta_,
               Function       f_oobag_eval,
-              const char&    type_oobag_eval_){
+              const char&    type_oobag_eval_,
+              const bool     verbose_progress){
 
 
  // convert inputs into arma objects
@@ -3353,11 +3354,17 @@ List orsf_fit(NumericMatrix& x,
   //
   // }
 
+  if(verbose_progress){
+   Rcout << "\r growing tree no. " << tree << " of " << n_tree;
+  }
+
+
   forest[tree] = ostree_fit(f_beta);
 
   // add 1 to tree here instead of end of loop
   // (more convenient to compute tree % oobag_eval_every)
   tree++;
+
 
   if(oobag_pred){
 
@@ -3399,7 +3406,9 @@ List orsf_fit(NumericMatrix& x,
 
  }
 
-
+ if(verbose_progress){
+  Rcout << std::endl;
+ }
 
  vec vimp(x_input.n_cols);